move mixin file, modify update poll refresh logic
